New character models and light fluid
Started cleaning up a character model to use in place of Ethan.
Added a script that makes lights produce a fluid that can be seen by the player and be used by the guards to adjust their eyesight.
The player can now scroll through all of the different fluid types.