HUD values separated from RaptorInteraction
Updated some stuff in Raptor Interaction
The code the controls the timers is separated into Timer.cs which is a component on the HUD now.
RaptorHUD is now a component on the HUD instead of the player.
The player prefab needs to be updated, but the HUD and Menus prefab should be updated already.