Fixed menus disappearing, door clinging, random traps
Changed menu cameras to clear depth only instead of nothing
Changed door tags to wood to prevent clinging
Made traps use random notoriety to determine when to spawn, door also use a random notoriety value for when to lock. This can be overridden by unchecking the random spawn/lock boxes on the object.