Made picking up stuff easier
Changes raycasts to sphere casts for detecting item prompts and using things.
Terminals and 2048 need to be tagged as "terminal"
There seems to be some issues with picking up keycards that are dropped by guards. Their tags are still set to enemy but I'm pretty sure that's not the option.